Backend/sicillian/requirements.txt
django
django-cors-headers
djangorestframework
PyJWT